resolved errors about constants and json sended to server

+++ tests/utils.go
@@ -17,7 +17,7 @@ const (
 	defSurname       = "TestSurname"
 	defEmail         = "TestEmail"
 	defClassroomName = "TestClassroom"
-	defClassroomYear = 1111
+	defClassroomYear = uint16(1111)
 )
 
 // logs in and gets the token
